Add files via upload
Project that I created with Damiano from the course. 

We used a dataset (hot or not) and facial recognition to extract features like eye size score, height to jaw ratio and multiple other metrics. Then we created a csv file that contained the values of the pictures including the values of multiple features. I trained a linear regression model to predict the attractiveness rating of woman, based on the scores that the facial metrics result in. We are going to build a neural net based on the same dataset to check if we can achieve better accuracy, as the model only achieves a mean error of 1.47, which, at this point is still disappointing to me.